SimulateSensor/requirements.txt
aiohttp
paho-mqtt